pyproject.toml
[tool.poetry]
name = "pytech"
version = "0.0.1"
description = ""
authors = ["Kyle Prestel <kprestel@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
pandas-datareader = "^0.7.0"
SQLAlchemy = "^1.3"
pandas-market-calendars = "^1.1"
psycopg2-binary = "^2.8"
"testing.postgresql" = "^1.3"
pytest-vcr = "^1.0"
pytest-docker-db = "^1.0"
hypothesis = "^4.31"
pymc3 = "^3.7"
pandas = "^0.24.0"

[tool.poetry.dev-dependencies]
pytest = "^4.0"
black = {version = "^18.3-alpha.0", allows-prereleases = true}
pytest-vcr = "^1.0"
